refactor: 유저 관리 로직 정리 (#19)
