+++ src/shared/components/admin/user-list/user-create-drawer.tsx
@@ -1,42 +1,16 @@
-import {
-  SignUpUserInfoDtoRole,
-  type CompanySignUpDTO,
-  type SignUpUserInfoDtoRole as DrawerRole,
-  type UserRole,
-  type UserRequestDTO,
-  useCreateCompany,
-  useUpdateUser,
-} from '@apis/telegro';
 import Icon from '@components/common/icon';
-import { toastError, toastSuccess } from '@components/common/toast/toast';
-import { useQueryClient } from '@tanstack/react-query';
 import { cn } from '@utils/cn';
-import { useEffect, useMemo, useState, type ChangeEvent } from 'react';
 import signupLogo from '/signup-logo.svg';
+import {
+  inputClass,
+  labelClass,
+  ROLE_OPTIONS,
+  STEP_FIELDS,
+} from './user-drawer.constants';
+import { useUserCreateDrawer } from './use-user-create-drawer';
+import type { UserDrawerInitialData } from './user-drawer.types';
 
-export type CreateCompanyForm = {
-  username: string;
-  userid: string;
-  password: string;
-  companyName: string;
-  phone: string;
-  email: string;
-  managerName: string;
-  managerPhone: string;
-  companyNumber: string;
-  companyType: string;
-  companyItem: string;
-  address: string;
-  zipCode: string;
-  addressDetail: string;
-  companyDescription: string;
-};
-
-export type UserDrawerInitialData = {
-  userId?: number;
-  role: DrawerRole;
-  form: CreateCompanyForm;
-};
+export type { CreateCompanyForm, UserDrawerInitialData } from './user-drawer.types';
 
 type UserCreateDrawerProps = {
   open: boolean;
@@ -45,467 +19,33 @@ type UserCreateDrawerProps = {
   onClose: () => void;
 };
 
-type StepField = {
-  key: keyof CreateCompanyForm;
-  label: string;
-  placeholder: string;
-  required?: boolean;
-  type?: 'text' | 'email' | 'password' | 'tel';
-  actionLabel?: string;
-};
-
-type DaumPostcodeData = {
-  zonecode: string;
-  address: string;
-  addressType: 'R' | 'J';
-  bname: string;
-  buildingName: string;
-  apartment: 'Y' | 'N';
-};
-
-declare global {
-  interface Window {
-    daum?: {
-      Postcode: new (options: {
-        oncomplete: (data: DaumPostcodeData) => void;
-      }) => {
-        open: () => void;
-      };
-    };
-  }
-}
-
-const POSTCODE_SCRIPT_ID = 'daum-postcode-script';
-const POSTCODE_SCRIPT_SRC =
-  'https://t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js';
-
-const ROLE_OPTIONS: Array<{ label: string; value: DrawerRole }> = [
-  { label: 'Member', value: SignUpUserInfoDtoRole.MEMBER },
-  { label: 'Dealer', value: SignUpUserInfoDtoRole.DEALER },
-  { label: 'Best', value: SignUpUserInfoDtoRole.BEST },
-  { label: 'Business', value: SignUpUserInfoDtoRole.BUSINESS },
-  { label: 'Admin', value: SignUpUserInfoDtoRole.ADMIN },
-];
-
-const INITIAL_FORM: CreateCompanyForm = {
-  username: '',
-  userid: '',
-  password: '',
-  companyName: '',
-  phone: '',
-  email: '',
-  managerName: '',
-  managerPhone: '',
-  companyNumber: '',
-  companyType: '',
-  companyItem: '',
-  address: '',
-  zipCode: '',
-  addressDetail: '',
-  companyDescription: '',
-};
-
-const STEP_FIELDS: Array<{
-  step: number;
-  title: string;
-  fields: StepField[];
-}> = [
-  {
-    step: 1,
-    title: '회원 유형 선택',
-    fields: [
-      {
-        key: 'username',
-        label: '회원명',
-        placeholder: '회원명을 입력해 주세요.',
-        required: true,
-      },
-      {
-        key: 'userid',
-        label: '아이디',
-        placeholder: '아이디를 입력해 주세요.',
-        required: true,
-      },
-      {
-        key: 'password',
-        label: '비밀번호',
-        placeholder: '비밀번호를 입력해 주세요.',
-        required: true,
-        type: 'password',
-      },
-      {
-        key: 'companyName',
-        label: '상호명',
-        placeholder: '상호명을 입력해 주세요.',
-        required: true,
-      },
-    ],
-  },
-  {
-    step: 2,
-    title: '연락처 정보',
-    fields: [
-      {
-        key: 'phone',
-        label: '전화번호',
-        placeholder: '전화번호를 입력해 주세요.',
-        required: true,
-        type: 'tel',
-      },
-      {
-        key: 'email',
-        label: '이메일(세금계산서용)',
-        placeholder: '이메일을 입력해 주세요.',
-        required: true,
-        type: 'email',
-      },
-      {
-        key: 'managerName',
-        label: '담당자 이름',
-        placeholder: '담당자 이름을 입력해 주세요.',
-        required: true,
-      },
-      {
-        key: 'managerPhone',
-        label: '담당자 전화번호',
-        placeholder: '담당자 전화번호를 입력해 주세요.',
-        required: true,
-        type: 'tel',
-      },
-    ],
-  },
-  {
-    step: 3,
-    title: '사업자 정보',
-    fields: [
-      {
-        key: 'companyNumber',
-        label: '사업자 번호',
-        placeholder: '사업자 번호를 입력해 주세요.',
-        required: true,
-      },
-      {
-        key: 'companyType',
-        label: '업태',
-        placeholder: '업태를 입력해 주세요.',
-        required: true,
-      },
-      {
-        key: 'companyItem',
-        label: '종목',
-        placeholder: '종목을 입력해 주세요.',
-        required: true,
-      },
-      {
-        key: 'address',
-        label: '주소',
-        placeholder: '주소를 검색해 주세요.',
-        required: true,
-        actionLabel: '주소 검색',
-      },
-      {
-        key: 'zipCode',
-        label: '우편번호',
-        placeholder: '우편번호',
-        required: true,
-      },
-      {
-        key: 'addressDetail',
-        label: '상세주소',
-        placeholder: '상세주소를 입력해 주세요.',
-        required: true,
-      },
-    ],
-  },
-  {
-    step: 4,
-    title: '메모',
-    fields: [
-      {
-        key: 'companyDescription',
-        label: '메모',
-        placeholder: '메모를 입력해 주세요.',
-      },
-    ],
-  },
-];
-
-const labelClass =
-  "font-['Pretendard',sans-serif] text-[1.25rem] font-semibold text-[#2B2B2B]";
-
-const inputClass =
-  "h-[5.6rem] w-full rounded-[1rem] border border-[#E9E9E9] bg-white px-[1.6rem] font-['Pretendard',sans-serif] text-[1.6rem] font-normal text-[#2B2B2B] outline-none transition placeholder:text-[#6D6D6D] focus:border-[#FFC633]";
-
-const buildRoadAddress = (data: DaumPostcodeData) => {
-  if (data.addressType !== 'R') {
-    return data.address;
-  }
-
-  const extras = [
-    data.bname,
-    data.apartment === 'Y' ? data.buildingName : '',
-  ].filter(Boolean);
-
-  return extras.length === 0 ? data.address : `${data.address} (${extras.join(', ')})`;
-};
-
-const getErrorMessage = (error: unknown, mode: 'create' | 'edit') => {
-  if (
-    typeof error === 'object' &&
-    error !== null &&
-    'response' in error &&
-    typeof error.response === 'object' &&
-    error.response !== null &&
-    'status' in error.response &&
-    error.response.status === 409
-  ) {
-    return '이미 사용 중인 상호명 혹은 ID입니다.';
-  }
-
-  if (
-    typeof error === 'object' &&
-    error !== null &&
-    'response' in error &&
-    typeof error.response === 'object' &&
-    error.response !== null &&
-    'data' in error.response &&
-    typeof error.response.data === 'object' &&
-    error.response.data !== null &&
-    'message' in error.response.data &&
-    typeof error.response.data.message === 'string'
-  ) {
-    return error.response.data.message;
-  }
-
-  return mode === 'edit'
-    ? '사용자 정보 수정에 실패했습니다.'
-    : '사용자 등록에 실패했습니다.';
-};
-
 const UserCreateDrawer = ({
   open,
   mode = 'create',
   initialData,
   onClose,
 }: UserCreateDrawerProps) => {
-  const queryClient = useQueryClient();
-  const createCompanyMutation = useCreateCompany();
-  const updateUserMutation = useUpdateUser();
-  const [currentStep, setCurrentStep] = useState(1);
-  const [selectedRole, setSelectedRole] = useState<DrawerRole>(
-    SignUpUserInfoDtoRole.MEMBER,
-  );
-  const [editingUserId, setEditingUserId] = useState<number | null>(null);
-  const [isPostcodeReady, setIsPostcodeReady] = useState(false);
-  const [form, setForm] = useState<CreateCompanyForm>(INITIAL_FORM);
-
-  const isEditMode = mode === 'edit';
-  const isSubmitting = createCompanyMutation.isPending || updateUserMutation.isPending;
-
-  const resetDrawer = () => {
-    setCurrentStep(1);
-    setSelectedRole(SignUpUserInfoDtoRole.MEMBER);
-    setEditingUserId(null);
-    setForm(INITIAL_FORM);
-  };
-
-  const handleFieldChange =
-    (key: keyof CreateCompanyForm) =>
-    (event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-      setForm((prev) => ({ ...prev, [key]: event.target.value }));
-    };
-
-  const handleAddressSearch = () => {
-    if (!window.daum?.Postcode) {
-      toastError('주소 검색을 불러오는 중입니다. 잠시 후 다시 시도해 주세요.');
-      return;
-    }
-
-    new window.daum.Postcode({
-      oncomplete: (data) => {
-        setForm((prev) => ({
-          ...prev,
-          zipCode: data.zonecode,
-          address: buildRoadAddress(data),
-        }));
-      },
-    }).open();
-  };
-
-  const validateStep = (step: number) => {
-    const stepConfig = STEP_FIELDS.find((item) => item.step === step);
-
-    if (!stepConfig) {
-      return true;
-    }
-
-    const hasEmptyRequiredField = stepConfig.fields.some((field) => {
-      if (isEditMode && field.key === 'password') {
-        return false;
-      }
-
-      return field.required && !form[field.key].trim();
-    });
-
-    if (hasEmptyRequiredField) {
-      toastError('모든 필수 항목을 입력해 주세요.');
-      return false;
-    }
-
-    return true;
-  };
-
-  const activeStep = useMemo(
-    () => STEP_FIELDS.find((item) => item.step === currentStep) ?? STEP_FIELDS[0],
-    [currentStep],
-  );
-
-  const isLastStep = currentStep === STEP_FIELDS.length;
-
-  const handleNextStep = () => {
-    if (!validateStep(currentStep)) {
-      return;
-    }
-
-    setCurrentStep((step) => Math.min(STEP_FIELDS.length, step + 1));
-  };
-
-  const handleSubmit = async () => {
-    if (!validateStep(1) || !validateStep(2) || !validateStep(3)) {
-      return;
-    }
-
-    try {
-      if (isEditMode) {
-        if (editingUserId == null) {
-          toastError('수정할 사용자 정보를 찾을 수 없습니다.');
-          return;
-        }
-
-        const payload: UserRequestDTO = {
-          user: {
-            username: form.username.trim(),
-            userId: form.userid.trim(),
-            ...(form.password.trim() ? { password: form.password } : {}),
-            role: selectedRole as UserRole,
-            phone: form.phone.trim(),
-            email: form.email.trim(),
-            address: form.address.trim(),
-            addressDetail: form.addressDetail.trim(),
-            zipCode: form.zipCode.trim(),
-          },
-          company: {
-            companyName: form.companyName.trim(),
-            managerName: form.managerName.trim(),
-            managerPhone: form.managerPhone.trim(),
-            companyNumber: form.companyNumber.trim(),
-            companyType: form.companyType.trim(),
-            companyItem: form.companyItem.trim(),
-            companyDescription: form.companyDescription.trim(),
-          },
-        };
-
-        await updateUserMutation.mutateAsync({ userId: editingUserId, data: payload });
-        toastSuccess('사용자 정보가 성공적으로 수정되었습니다.');
-      } else {
-        const payload: CompanySignUpDTO = {
-          signUpUserInfoDto: {
-            userid: form.userid.trim(),
-            username: form.username.trim(),
-            password: form.password,
-            phone: form.phone.trim(),
-            email: form.email.trim(),
-            role: selectedRole,
-            address: form.address.trim(),
-            addressDetail: form.addressDetail.trim(),
-            zipCode: form.zipCode.trim(),
-          },
-          company: {
-            managerName: form.managerName.trim(),
-            managerPhone: form.managerPhone.trim(),
-            companyName: form.companyName.trim(),
-            companyNumber: form.companyNumber.trim(),
-            companyType: form.companyType.trim(),
-            companyItem: form.companyItem.trim(),
-            companyDescription: form.companyDescription.trim(),
-          },
-        };
-
-        await createCompanyMutation.mutateAsync({ data: payload });
-        toastSuccess('사용자 정보가 성공적으로 등록되었습니다.');
-      }
-
-      await queryClient.invalidateQueries({ queryKey: ['/api/users'] });
-      resetDrawer();
-      onClose();
-    } catch (error) {
-      toastError(getErrorMessage(error, mode));
-    }
-  };
-
-  useEffect(() => {
-    if (window.daum?.Postcode) {
-      setIsPostcodeReady(true);
-      return;
-    }
-
-    const existingScript = document.getElementById(
-      POSTCODE_SCRIPT_ID,
-    ) as HTMLScriptElement | null;
-    const handleLoad = () => setIsPostcodeReady(true);
-
-    if (existingScript) {
-      existingScript.addEventListener('load', handleLoad, { once: true });
-      return () => existingScript.removeEventListener('load', handleLoad);
-    }
-
-    const script = document.createElement('script');
-    script.id = POSTCODE_SCRIPT_ID;
-    script.src = POSTCODE_SCRIPT_SRC;
-    script.async = true;
-    script.addEventListener('load', handleLoad, { once: true });
-    document.body.appendChild(script);
-
-    return () => script.removeEventListener('load', handleLoad);
-  }, []);
-
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-
-    const previousOverflow = document.body.style.overflow;
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        onClose();
-      }
-    };
-
-    document.body.style.overflow = 'hidden';
-    window.addEventListener('keydown', handleKeyDown);
-
-    return () => {
-      document.body.style.overflow = previousOverflow;
-      window.removeEventListener('keydown', handleKeyDown);
-    };
-  }, [open, onClose]);
-
-  useEffect(() => {
-    if (!open) {
-      resetDrawer();
-      return;
-    }
-
-    if (isEditMode && initialData) {
-      setCurrentStep(1);
-      setEditingUserId(initialData.userId ?? null);
-      setSelectedRole(initialData.role);
-      setForm(initialData.form);
-      return;
-    }
-
-    resetDrawer();
-  }, [initialData, isEditMode, open]);
+  const {
+    activeStep,
+    currentStep,
+    form,
+    handleAddressSearch,
+    handleFieldChange,
+    handleNextStep,
+    handleSubmit,
+    isEditMode,
+    isLastStep,
+    isPostcodeReady,
+    isSubmitting,
+    selectedRole,
+    setCurrentStep,
+    setSelectedRole,
+  } = useUserCreateDrawer({
+    open,
+    mode,
+    initialData,
+    onClose,
+  });
 
   return (
     <>
